backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib
python-jose
pymysql
bcrypt